+++ modi_plus/module/input_module/button.py
@@ -1,7 +1,7 @@
 """Button module."""
 
 import struct
-from modi_plus.module.input_module.input_module import InputModule
+from modi_plus.module.module import InputModule
 
 
 class Button(InputModule):
