--- modi_plus/module/input_module/input_module.py
@@ -1,5 +1,0 @@
-from modi_plus.module.module import Module
-
-
-class InputModule(Module):
-    pass
