+++ modi_plus/module/input_module/tof.py
@@ -1,7 +1,7 @@
 """Tof module."""
 
 import struct
-from modi_plus.module.input_module.input_module import InputModule
+from modi_plus.module.module import InputModule
 
 
 class Tof(InputModule):
