+++ modi_plus/module/output_module/led.py
@@ -3,7 +3,7 @@
 import struct
 
 from typing import Tuple
-from modi_plus.module.output_module.output_module import OutputModule
+from modi_plus.module.module import OutputModule
 
 
 class Led(OutputModule):
