+++ modi_plus/module/setup_module/network.py
@@ -2,7 +2,7 @@
 
 import time
 import struct
-from modi_plus.module.setup_module.setup_module import SetupModule
+from modi_plus.module.module import SetupModule
 
 
 class Network(SetupModule):
